requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
requests==2.32.3
python-dotenv==1.0.1
pillow==10.4.0
pytesseract==0.3.13
pdfminer.six==20240706
openai==1.51.2
pypdfium2==4.30.0
httpx==0.27.2